remove chat window tabs + remove tooltip + code cleanup
