+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/ChatWidgetHUD/ChatEntry.cs
@@ -4,45 +4,6 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-public static class RectTransformExtensions
-{
-    public static int CountCornersVisibleFrom(this RectTransform rectTransform, RectTransform viewport)
-    {
-        Vector3[] viewCorners = new Vector3[4];
-        viewport.GetWorldCorners(viewCorners);
-
-        Vector2 size = new Vector2(viewport.rect.size.x * viewport.lossyScale.x, viewport.rect.size.y * viewport.lossyScale.y);
-        Rect screenBounds = new Rect(viewCorners[0], size); // Screen space bounds (assumes camera renders across the entire screen)
-
-        Vector3[] objectCorners = new Vector3[4];
-        rectTransform.GetWorldCorners(objectCorners);
-
-        int visibleCorners = 0;
-
-        for (var i = 0; i < viewCorners.Length; i++)
-        {
-            if (i != viewCorners.Length - 1)
-                Debug.DrawLine(viewCorners[i], viewCorners[i + 1], Color.blue, 1.0f);
-        }
-
-        for (var i = 0; i < objectCorners.Length; i++) // For each corner in rectTransform
-        {
-            if (screenBounds.Contains(objectCorners[i])) // If the corner is inside the screen
-            {
-                visibleCorners++;
-            }
-        }
-
-        for (var i = 0; i < objectCorners.Length; i++) // For each corner in rectTransform
-        {
-            if (i != objectCorners.Length - 1)
-                Debug.DrawLine(objectCorners[i], objectCorners[i + 1], visibleCorners > 0 ? Color.green : Color.red, 1.0f);
-        }
-
-        return visibleCorners;
-    }
-}
-
 public class ChatEntry : MonoBehaviour
 {
     public struct Model
@@ -62,6 +23,9 @@ public class ChatEntry : MonoBehaviour
 
         public ulong timestamp;
     }
+
+    [SerializeField] internal float timeToFade = 10;
+    [SerializeField] internal float fadeDuration = 5;
 
     [SerializeField] internal TextMeshProUGUI username;
     [SerializeField] internal TextMeshProUGUI body;
@@ -115,9 +79,6 @@ public class ChatEntry : MonoBehaviour
 
     [SerializeField] CanvasGroup group;
 
-    const float TIME_TO_FADE = 10;
-    const float FADE_DURATION = 5;
-
     public void SetFadeout(bool enabled)
     {
         if (!enabled)
@@ -132,13 +93,13 @@ public class ChatEntry : MonoBehaviour
 
     private void Update()
     {
-        double fadeTime = (double)(model.timestamp / 1000.0) + TIME_TO_FADE;
+        double fadeTime = (double)(model.timestamp / 1000.0) + timeToFade;
         double currentTime = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds() / 1000.0;
 
         if (currentTime > fadeTime)
         {
             double timeSinceFadeTime = currentTime - fadeTime;
-            group.alpha = Mathf.Clamp01(1 - (float)(timeSinceFadeTime / FADE_DURATION));
+            group.alpha = Mathf.Clamp01(1 - (float)(timeSinceFadeTime / fadeDuration));
         }
         else
         {
