--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/ChatWidgetHUD/ExtTest.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-public class ExtTest : MonoBehaviour
-{
-    public RectTransform viewport;
-    void Update()
-    {
-        (transform as RectTransform).CountCornersVisibleFrom(viewport);
-    }
-}
