+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/TaskbarHUD/TaskbarHUDView.cs
@@ -10,8 +10,6 @@ public class TaskbarHUDView : MonoBehaviour
     public Button chatButton;
     public Button friendsButton;
 
-    public GameObject chatTooltip;
-
     internal TaskbarHUDController controller;
 
     internal static TaskbarHUDView Create(TaskbarHUDController controller)
@@ -24,25 +22,6 @@ public class TaskbarHUDView : MonoBehaviour
     public void Initialize(TaskbarHUDController controller)
     {
         this.controller = controller;
-        chatTooltip.SetActive(false);
-        CommonScriptableObjects.rendererState.OnChange -= RendererState_OnChange;
-        CommonScriptableObjects.rendererState.OnChange += RendererState_OnChange;
-    }
-
-    private void OnDestroy()
-    {
-        CommonScriptableObjects.rendererState.OnChange -= RendererState_OnChange;
-    }
-
-    private void RendererState_OnChange(bool current, bool previous)
-    {
-        if (current == previous)
-            return;
-
-        if (current && !controller.alreadyToggledOnForFirstTime)
-        {
-            chatTooltip.SetActive(true);
-        }
     }
 
     internal void OnAddChatWindow(UnityAction onToggle)
@@ -60,12 +39,6 @@ public class TaskbarHUDView : MonoBehaviour
     public void SetVisibility(bool visible)
     {
         gameObject.SetActive(visible);
-    }
-
-    public void OnToggleForFirstTime()
-    {
-        //TODO(Brian): Toggle an animator trigger/bool instead of doing this.
-        chatTooltip.SetActive(false);
     }
 
     private void Update()
