+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/Tests/WorldChatWindowHUDShould.cs
@@ -4,6 +4,7 @@ using NUnit.Framework;
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using UnityEngine;
 using UnityEngine.TestTools;
 
 class ChatController_Mock : IChatController
@@ -75,81 +76,6 @@ public class WorldChatWindowHUDShould : TestsBase
         yield break;
     }
 
-    [Test]
-    public void TabsWorkCorrectly()
-    {
-
-        var messages = new ChatMessage[]
-        {
-            new ChatMessage()
-            {
-                messageType = ChatMessage.Type.PUBLIC,
-                body = "test message 1",
-                sender = ownProfileModel.userId,
-                recipient = testProfileModel.userId
-            },
-            new ChatMessage()
-            {
-                messageType = ChatMessage.Type.PUBLIC,
-                body = "test message 2",
-                sender = ownProfileModel.userId,
-                recipient = testProfileModel.userId
-            },
-            new ChatMessage()
-            {
-                messageType = ChatMessage.Type.PRIVATE,
-                body = "test message 3",
-                sender = ownProfileModel.userId,
-                recipient = testProfileModel.userId
-            },
-            new ChatMessage()
-            {
-                messageType = ChatMessage.Type.PRIVATE,
-                body = "test message 4",
-                sender = ownProfileModel.userId,
-                recipient = testProfileModel.userId
-            },
-        };
-
-        foreach (var msg in messages)
-        {
-            chatController.RaiseAddMessage(msg);
-        }
-
-        var expectedBodyMessages = new string[]
-        {
-            "<b>NO_USER:</b> test message 1",
-            "<b>NO_USER:</b> test message 2",
-            "<b>[To TEST_USER]:</b> test message 3",
-            "<b>[To TEST_USER]:</b> test message 4"
-        };
-
-        Assert.AreEqual(4, controller.view.chatHudView.entries.Count);
-
-        for (int i = 0; i < controller.view.chatHudView.entries.Count; i++)
-        {
-            ChatEntry entry = controller.view.chatHudView.entries[i];
-            Assert.AreEqual(expectedBodyMessages[i], entry.body.text);
-        }
-
-        controller.view.pmFilterButton.onClick.Invoke();
-
-        expectedBodyMessages = new string[]
-        {
-            "<b>[To TEST_USER]:</b> test message 3",
-            "<b>[To TEST_USER]:</b> test message 4"
-        };
-
-        Assert.AreEqual(2, controller.view.chatHudView.entries.Count);
-
-        for (int i = 0; i < controller.view.chatHudView.entries.Count; i++)
-        {
-            ChatEntry entry = controller.view.chatHudView.entries[i];
-            Assert.AreEqual(expectedBodyMessages[i], entry.body.text);
-        }
-
-        controller.view.worldFilterButton.onClick.Invoke();
-    }
 
     [Test]
     public void HandlePrivateMessagesProperly()
@@ -233,10 +159,14 @@ public class WorldChatWindowHUDShould : TestsBase
 
         WebInterface.OnMessageFromEngine += messageCallback;
         controller.resetInputFieldOnSubmit = false;
-        controller.SendChatMessage("test message");
-        Assert.IsTrue(messageWasSent);
-        Assert.AreEqual("", controller.view.chatHudView.inputField.text);
+        controller.view.chatHudView.inputField.onSubmit.Invoke("test message");
+
+        Debug.Log("text = " + controller.view.chatHudView.inputField.text);
+
         WebInterface.OnMessageFromEngine -= messageCallback;
+        yield return null;
+        yield return null;
+        yield return null;
         yield break;
     }
