+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/WorldChatWindowHUDController.cs
@@ -116,6 +116,8 @@ public class WorldChatWindowHUDController : IHUD
             sender = UserProfile.GetOwnUserProfile().userId,
         };
 
+        view.chatHudView.inputField.text = "sarlanga";
+
         WebInterface.SendChatMessage(data);
     }
