+++ unity-client/Assets/Scripts/MainScripts/DCL/Helpers/Utils/RectTransformExtensions.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 960c7de2a45a8864a84c55656af01105
+guid: a7c878f914298824cacfb1fea26279f6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
